Began adding finishing touches to the player refactor, added pushable objects, updated networking, added UI hints
Almost finished the player refactor, just need to add the ability to interact with switches and finish refactoring the player input system.
Pushable objects now work mostly as intended, only needs fixing on slopes.
The refactor so far works completely with the networking, and it much smoother.
Added a controls page and pop-up UI hints when standing next to an interactable to prompt you to use the correct key.