Multiple menu changes
Changes to the main menu: Added a custom splash screen, added a cinematic panning up to the first level, added effects to all buttons to they are now standardised.
Changes to the first level: Added a cinematic panning down from from the sky to the level, fixed bugs with not being able to quit the game.