Added slingshot ability and ground pound ability
Added slingshot ability: Holding right click when close to specific objects enters aim mode. Releasing right click will launch the player in the direction of the mouth. Tied to the Water element.
Added ground pound ability: Pressing S while in the air will cause the player to fly directly downwards in a "ground pound" state, which can be used to trigger things below it. Tied to the Earth element.